backend/requirements.txt
fastapi[standard]
uvicorn
openai-whisper
python-multipart
soundfile
pydub
pikepdf
